Add Inscripcion, EmailService, and related DTOs; implement FindByEmailAsync in IAlumnoRepository

+++ PortalGalaxy/src/PortalGalaxy.Services/Implementaciones/TallerService.cs
@@ -1,4 +1,4 @@
-using AutoMapper;
+﻿using AutoMapper;
 using Microsoft.Extensions.Logging;
 using PortalGalaxy.Common.Request;
 using PortalGalaxy.Common.Response;
@@ -24,24 +24,71 @@ public class TallerService : ITallerService
         _fileUploader = fileUploader;
     }
 
-    public async Task<PaginationResponse<TallerDtoResponse>> ListAsync(BusquedaTallerRequest request)
+    public async Task<PaginationResponse<InscritosPorTallerDtoResponse>> ListAsync(BusquedaInscritosPorTallerRequest request)
     {
-        var response = new PaginationResponse<TallerDtoResponse>();
+        var response = new PaginationResponse<InscritosPorTallerDtoResponse>();
 
         try
         {
-            var (lista, total) = await _repository.ListAsync(request.Nombre, request.Categoria, 
-                request.Situacion, request.PageNumber, request.PageSize);
-            response.Data = _mapper.Map<ICollection<TallerDtoResponse>>(lista);
-            response.TotalPages = Helper.GetTotalPages(total, request.PageSize); // Calcular el total de páginas
-            response.TotalCount = total;
+            // Codigo
+            var tupla = await _repository.ListAsync(request.InstructorId, request.Taller, request.Situacion,
+                request.FechaInicio, request.FechaFin, request.PageNumber, request.PageSize);
+
+            response.Data = _mapper.Map<ICollection<InscritosPorTallerDtoResponse>>(tupla.Colecction);
+            response.TotalPages = Helper.GetTotalPages(tupla.Total, request.PageSize);
             response.Success = true;
         }
         catch (Exception ex)
         {
-            _logger.LogError(ex, "Error in TallerService.ListAsync");
-            response.Success = false;
-            response.ErrorMessage = "Error al listar los talleres.";
+            response.ErrorMessage = "Error al listar los inscritos por taller";
+            _logger.LogError(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
+        }
+
+        return response;
+    }
+
+    public async Task<BaseResponse<ICollection<TallerSimpleDtoResponse>>> ListSimpleAsync()
+    {
+        var response = new BaseResponse<ICollection<TallerSimpleDtoResponse>>();
+
+        try
+        {
+            response.Data = await _repository.ListAsync(
+                predicate: x => x.Situacion == SituacionTaller.Aperturada || x.Situacion == SituacionTaller.Por_Aperturar,
+                selector: x => new TallerSimpleDtoResponse
+                {
+                    Id = x.Id,
+                    Nombre = x.Nombre
+                });
+
+            response.Success = true;
+        }
+        catch (Exception ex)
+        {
+            response.ErrorMessage = "Error al cargar los talleres";
+            _logger.LogCritical(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
+        }
+
+        return response;
+    }
+
+    public async Task<PaginationResponse<TallerHomeDtoResponse>> ListarTalleresHomeAsync(BusquedaTallerHomeRequest request)
+    {
+        var response = new PaginationResponse<TallerHomeDtoResponse>();
+
+        try
+        {
+            var tupla = await _repository.ListarTalleresHomeAsync(request.Nombre, request.InstructorId,
+                request.FechaInicio, request.FechaFin, request.PageNumber, request.PageSize);
+
+            response.Data = _mapper.Map<ICollection<TallerHomeDtoResponse>>(tupla.Collection);
+            response.TotalPages = Helper.GetTotalPages(tupla.Total, request.PageSize);
+            response.Success = true;
+        }
+        catch (Exception ex)
+        {
+            response.ErrorMessage = "Error al listar los talleres";
+            _logger.LogError(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
         }
 
         return response;
@@ -62,8 +109,8 @@ public class TallerService : ITallerService
         }
         catch (Exception ex)
         {
-            response.ErrorMessage = "Error al agregar el taller";
-            _logger.LogCritical(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
+            response.ErrorMessage = "Error al agregar un Taller";
+            _logger.LogError(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
         }
         return response;
     }
@@ -76,16 +123,76 @@ public class TallerService : ITallerService
             var entity = await _repository.FindAsync(id);
             if (entity == null)
             {
-                response.ErrorMessage = "Taller no encontrado";
+                response.ErrorMessage = "No se encontró el Taller";
                 return response;
             }
+
             response.Data = _mapper.Map<TallerDtoRequest>(entity);
             response.Success = true;
         }
         catch (Exception ex)
         {
-            response.ErrorMessage = "Error al buscar el taller";
-            _logger.LogCritical(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
+            response.ErrorMessage = "Error al buscar un Taller";
+            _logger.LogError(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
+        }
+        return response;
+    }
+
+    public async Task<BaseResponse<TallerHomeDtoResponse>> GetTallerHomeAsync(int id)
+    {
+        var response = new BaseResponse<TallerHomeDtoResponse>();
+        try
+        {
+            var entity = await _repository.ObtenerTallerHomeAsync(id);
+            if (entity == null)
+            {
+                response.ErrorMessage = "No se encontró el Taller";
+                return response;
+            }
+
+            response.Data = _mapper.Map<TallerHomeDtoResponse>(entity);
+            response.Success = true;
+        }
+        catch (Exception ex)
+        {
+            response.ErrorMessage = "Error al buscar un Taller";
+            _logger.LogError(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
+        }
+        return response;
+    }
+
+    public async Task<BaseResponse<ICollection<TalleresPorMesDto>>> ReporteTalleresPorMes(int anio)
+    {
+        var response = new BaseResponse<ICollection<TalleresPorMesDto>>();
+        try
+        {
+            var entity = await _repository.ListarTalleresPorMesAsync(anio);
+
+            response.Data = _mapper.Map<ICollection<TalleresPorMesDto>>(entity);
+            response.Success = true;
+        }
+        catch (Exception ex)
+        {
+            response.ErrorMessage = "Error al buscar un Taller";
+            _logger.LogError(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
+        }
+        return response;
+    }
+
+    public async Task<BaseResponse<ICollection<TalleresPorInstructorDto>>> ReporteTalleresPorInstructor(int anio)
+    {
+        var response = new BaseResponse<ICollection<TalleresPorInstructorDto>>();
+        try
+        {
+            var entity = await _repository.ListarTalleresPorInstructorAsync(anio);
+
+            response.Data = _mapper.Map<ICollection<TalleresPorInstructorDto>>(entity);
+            response.Success = true;
+        }
+        catch (Exception ex)
+        {
+            response.ErrorMessage = "Error al buscar un Taller";
+            _logger.LogError(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
         }
         return response;
     }
@@ -98,17 +205,18 @@ public class TallerService : ITallerService
             var entity = await _repository.FindAsync(id);
             if (entity == null)
             {
-                response.ErrorMessage = "Taller no encontrado";
+                response.ErrorMessage = "No se encontró el Taller";
                 return response;
             }
-            _mapper.Map(request, entity);
 
-            if (!string.IsNullOrEmpty(request.PortadaBase64) && !string.IsNullOrEmpty(request.PortadaFileName))
+            _mapper.Map(request, entity);
+            
+            if (request.PortadaBase64 != null)
             {
                 entity.PortadaUrl = await _fileUploader.UploadFileAsync(request.PortadaBase64, request.PortadaFileName);
             }
-
-            if (!string.IsNullOrEmpty(request.TemarioBase64) && !string.IsNullOrEmpty(request.TemarioFileName))
+            
+            if (request.TemarioBase64 != null)
             {
                 entity.TemarioUrl = await _fileUploader.UploadFileAsync(request.TemarioBase64, request.TemarioFileName);
             }
@@ -118,8 +226,8 @@ public class TallerService : ITallerService
         }
         catch (Exception ex)
         {
-            response.ErrorMessage = "Error al actualizar el taller";
-            _logger.LogCritical(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
+            response.ErrorMessage = "Error al actualizar un Taller";
+            _logger.LogError(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
         }
         return response;
     }
@@ -129,13 +237,40 @@ public class TallerService : ITallerService
         var response = new BaseResponse();
         try
         {
+            var entity = await _repository.FindAsync(id);
+            if (entity == null)
+            {
+                response.ErrorMessage = "No se encontró el Taller";
+                return response;
+            }
+
             await _repository.DeleteAsync(id);
             response.Success = true;
         }
         catch (Exception ex)
         {
-            response.ErrorMessage = "Error al eliminar el taller";
-            _logger.LogCritical(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
+            response.ErrorMessage = "Error al eliminar un Taller";
+            _logger.LogError(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
+        }
+        return response;
+    }
+
+    public async Task<PaginationResponse<TallerDtoResponse>> ListAsync(BusquedaTallerRequest request)
+    {
+        var response = new PaginationResponse<TallerDtoResponse>();
+        try
+        {
+            var tupla = await _repository.ListarTalleresAsync(request.Nombre, request.Categoria, request.Situacion, request.PageNumber, request.PageSize);
+
+            response.Data = _mapper.Map<ICollection<TallerDtoResponse>>(tupla.Collection);
+            response.TotalPages = Helper.GetTotalPages(tupla.Total, request.PageSize);
+
+            response.Success = true;
+        }
+        catch (Exception ex)
+        {
+            response.ErrorMessage = "Error al listar los Talleres";
+            _logger.LogError(ex, "{ErrorMessage} {Message}", response.ErrorMessage, ex.Message);
         }
         return response;
     }
